+++ packages/internet_file/lib/src/internet_file.dart
@@ -4,13 +4,20 @@ import 'dart:typed_data';
 import 'package:http/http.dart' as http;
 import 'package:internet_file/src/storage.dart';
 
+@Deprecated('Usage `progress` instead')
 typedef InternetFileProcess = void Function(double percentage);
+typedef InternetFileProgress = void Function(
+  int receivedLength,
+  int contentLength,
+);
 
 class InternetFile {
   static Future<Uint8List> get(
     String url, {
     Map<String, String>? headers,
-    InternetFileProcess? process,
+    // ignore: deprecated_member_use_from_same_package
+    @Deprecated('Usage `progress` instead') InternetFileProcess? process,
+    InternetFileProgress? progress,
     InternetFileStorage? storage,
     InternetFileStorageAdditional storageAdditional = const {},
     bool force = false,
@@ -25,8 +32,8 @@ class InternetFile {
     }
     final response = httpClient.send(request);
 
-    List<List<int>> chunks = [];
-    int downloaded = 0;
+    List<int> bytesList = [];
+    int receivedLength = 0;
 
     if (storage != null) {
       final localResult = await storage.findExist(url, storageAdditional);
@@ -38,34 +45,26 @@ class InternetFile {
     response.asStream().listen((http.StreamedResponse request) {
       request.stream.listen(
         (List<int> chunk) {
-          final contentLength = request.contentLength ?? 0;
-          final percentage = downloaded / contentLength * 100;
-          // Display percentage of completion
+          receivedLength += chunk.length;
+          final contentLength = request.contentLength ?? receivedLength;
+          final percentage = receivedLength / contentLength * 100;
+
           if (debug) {
-            print('downloadPercentage: $percentage');
+            print('download progress: $receivedLength of '
+                '$contentLength ($percentage%)');
           }
           process?.call(percentage);
+          progress?.call(receivedLength, contentLength);
 
-          chunks.add(chunk);
-          downloaded += chunk.length;
+          bytesList.addAll(chunk);
         },
         onDone: () {
-          final contentLength = request.contentLength ?? 0;
-          final percentage = downloaded / contentLength * 100;
+          final bytes = Uint8List.fromList(bytesList);
 
-          if (debug) {
-            print('downloadPercentage: $percentage');
-          }
-          process?.call(percentage);
-
-          final Uint8List bytes = Uint8List(contentLength);
-          int offset = 0;
-
-          for (List<int> chunk in chunks) {
-            bytes.setRange(offset, offset + chunk.length, chunk);
-            offset += chunk.length;
-          }
           if (storage != null) {
+            if (debug) {
+              print('Save file in storage...');
+            }
             storage.save(url, storageAdditional, bytes);
           }
           completer.complete(bytes);